Refactor: Component 별로 폴더 생성, 폴더 위치 조정
