+++ src/pages/sunghyunKim/Main/Feeds/Feeds.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Feed from '../Feed/Feed';
+import Feed from './Feed/Feed';
 import './Feeds.scss';
 function Feeds() {
   const feed1Comment = [
