+++ src/pages/sunghyunKim/Main/MainKim.js
@@ -2,11 +2,8 @@ import React from 'react';
 import './MainKim.scss';
 import { useState } from 'react';
 import Nav from '../../../components/Nav/Nav';
-import Footer from './Footer/Footer';
-import Recommends from './Recommends/Recommends';
-import Stories from './Stories/Stories';
-import Myprofile from './Myprofile/Myprofile';
 import Feeds from './Feeds/Feeds';
+import MainRight from './MainRight/MainRight';
 
 function MainKim() {
   const [currentPopup, setCurrentPopup] = useState('');
@@ -20,12 +17,7 @@ function MainKim() {
       <Nav setCurrentPopup={setCurrentPopup} currentPopup={currentPopup} />
       <main className="main">
         <Feeds />
-        <section className="main-right">
-          <Myprofile />
-          <Stories />
-          <Recommends />
-          <Footer />
-        </section>
+        <MainRight />
       </main>
     </section>
   );
