+++ src/pages/sunghyunKim/Main/MainRight/Myprofile/Myprofile.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/common.scss';
+@import '../../../../../styles/common.scss';
 
 .myprofile {
   display: flex;
