+++ src/pages/sunghyunKim/Main/MainRight/Recommends/Recommend/Recommend.scss
@@ -1,0 +1,37 @@
+@import '../../../../../../styles/common.scss';
+.recommend {
+  display: flex;
+  justify-content: space-between;
+  margin: 7px 0;
+
+  &__main {
+    display: flex;
+
+    img {
+      object-fit: cover;
+    }
+  }
+
+  & > button {
+    font-size: 13px;
+    font-weight: 600;
+  }
+
+  &__info {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+    font-size: 13px;
+    margin-left: 5px;
+
+    & > span:first-child {
+      color: $color-black;
+      font-weight: 700;
+    }
+
+    & > span:last-child {
+      margin-top: 2px;
+      color: $color-gray;
+    }
+  }
+}
